enhancement: Implements the touch interpolation action as well as the mobile header animation store in animation lib
